Refactor routers and handlers
