+++ controller/brdownload_handler.go
@@ -1,0 +1,16 @@
+package controller
+
+import (
+	"net/http"
+
+	"github.com/julienschmidt/httprouter"
+)
+
+// FileBRDownloadHandler is to handle downloading files in breakpoint-resumed mode
+func FileBRDownloadHandler(w http.ResponseWriter, r *http.Request, ps httprouter.Params) {
+	// Parse the http request
+
+	// Business logic
+
+	// Return the http response
+}
