--- controller/index_handler.go
@@ -1,21 +1,0 @@
-package controller
-
-import (
-	"io/ioutil"
-	"log"
-	"net/http"
-
-	"github.com/julienschmidt/httprouter"
-)
-
-// IndexHandler is to handle directing to index.html
-func IndexHandler(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
-	data, err := ioutil.ReadFile("./static/view/signup.html")
-	if err != nil {
-		log.Fatal("Failed to read index.html, err: \n" + err.Error())
-
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-	w.Write(data)
-}
