+++ main.go
@@ -19,17 +19,22 @@ func main() {
 	router.GET("/", controller.IndexHandler)
 
 	// Handle file operations
+	router.POST("/file/upload", controller.FileUploadHandler)
+	router.GET("/file/query/filesha1", controller.SingleFileQueryHandler)
+	router.GET("/file/query/limitcount", controller.BatchFilesQueryHandler)
+	router.GET("/file/upload", controller.FileDownloadHandler)
+	router.PUT("/file/update", controller.FileUpdateHandler)
+	router.DELETE("/file/delete", controller.FileDeleteHandler)
+
+	router.POST("/file/ftupload", controller.FileFastUploadHandler)
 	router.POST("/file/mpupload/init", controller.FileMPUploadInitHandler)
 	router.POST("/file/mpupload/part", controller.FileMPUploadPartHandler)
 	router.POST("/file/mpupload/finish", controller.FileMPUploadFinishHandler)
 	router.POST("/file/mpupload/cancel", controller.FileMPUploadCancelHandler)
 	router.POST("/file/mpupload/status", controller.FileMPUploadStatusHandler)
-	router.POST("/file/upload", controller.FileUploadHandler)
-	router.GET("/file/query/filesha1", controller.SingleFileQueryHandler)
-	router.GET("/file/query/limitcount", controller.BatchFilesQueryHandler)
-	router.GET("/file/download", controller.FileDownloadHandler)
-	router.PUT("/file/update", controller.FileUpdateHandler)
-	router.DELETE("/file/delete", controller.FileDeleteHandler)
+	router.GET("/file/attachment", controller.FileDownloadAttachmentHandler)
+	router.GET("/file/url", controller.FileDownloadURLHandler)
+	router.GET("/file/brdownload", controller.FileBRDownloadHandler)
 
 	// Handle user operations
 	router.POST("/user/signup", controller.UserSignUpHandler)
